Rework AI summary: rename module, reactive availability, prompt and UX fixes
Rename core:app-ai-description to core:ai-insights and split it into a
shared ai/ engine layer plus an appdescription/ subpackage, so future
on-device AI features can reuse the engine without duplicating plumbing.

- Inject GenerativeModel via Hilt instead of instantiating it inline.
- Replace collect+var download tracking with firstOrNull.
- Expose AppAiDescriptionRepository.availability as a shared, hot
  StateFlow<AiAvailability> (Available/Downloadable/Downloading/Unavailable)
  instead of a one-shot suspend check, so every AiSummaryViewModel reacts
  live instead of polling.
- Make downloadModel() fire-and-forget and run in the repository's appScope
  (not viewModelScope), so navigating away no longer cancels an in-progress
  download; concurrent callers join the same in-flight download.
- Hash the fully-built prompt instead of a hand-maintained PROMPT_VERSION,
  so prompt/schema changes self-invalidate the cache.
- Collapse short/long description into a single description field with a
  purpose-focused prompt (permissions as supporting context, not a forced
  separate technical sentence); instruct the model to avoid opening with
  the app name or "This application"/"This app".
- Drop activities/services/receivers from the prompt context - permissions
  are the real signal and release-build class names are usually obfuscated.
- Only cache AppReference.InstalledPackage results; APK-file analysis is
  generated fresh each time to avoid colliding with the installed app's
  cached row (both share the same packageName-keyed cache row).
- Move AiSummaryCard/ViewModel into components/aisummary/, split
  AiSummaryState/AiSummaryAction out of the ViewModel, add a download
  action, a 2-line shimmer while generating, smaller "hint" typography,
  and animateContentSize() for smooth resize between states.
- Parameterize Modifier.shimmer()/SkeletonBox() colors so the AI card's
  shimmer can use the purple accent instead of the neutral default.
- Align AiSummaryViewModel.state and the repository's availability with
  this codebase's explicit-backing-field StateFlow convention.
