+++ core/ai-insights/CLAUDE.md
