+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/ai/AiAvailability.kt
@@ -1,0 +1,8 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.ai
+
+enum class AiAvailability {
+    Available,
+    Downloadable,
+    Downloading,
+    Unavailable,
+}
