+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/ai/OnDeviceAiEngine.kt
@@ -1,0 +1,9 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.ai
+
+internal interface OnDeviceAiEngine {
+    suspend fun checkAvailability(): AiAvailability
+
+    suspend fun downloadModel(): Boolean
+
+    suspend fun runPrompt(prompt: String): String?
+}
