+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/appdescription/AppAiDescription.kt
@@ -1,0 +1,3 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.appdescription
+
+data class AppAiDescription(val description: String)
