+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/appdescription/cache/AppAiDescriptionCache.kt
@@ -1,0 +1,13 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.cache
+
+import sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.AppAiDescription
+
+internal interface AppAiDescriptionCache {
+    suspend fun get(packageName: String, inputHash: String): AppAiDescription?
+
+    suspend fun save(
+        packageName: String,
+        inputHash: String,
+        description: AppAiDescription,
+    )
+}
