+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/appdescription/cache/AppAiDescriptionDatabase.kt
@@ -1,0 +1,9 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.cache
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = [AppAiDescriptionEntity::class], version = 3, exportSchema = false)
+internal abstract class AppAiDescriptionDatabase : RoomDatabase() {
+    abstract fun appAiDescriptionDao(): AppAiDescriptionDao
+}
