+++ core/ai-insights/src/main/kotlin/sk/styk/martin/apkanalyzer/core/aiinsights/appdescription/generation/AiDescriptionGenerator.kt
@@ -1,0 +1,15 @@
+package sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.generation
+
+import sk.styk.martin.apkanalyzer.core.aiinsights.ai.AiAvailability
+import sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.AppAiContext
+import sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.AppAiDescription
+
+internal interface AiDescriptionGenerator {
+    suspend fun checkAvailability(): AiAvailability
+
+    suspend fun downloadModel(): Boolean
+
+    fun inputHash(context: AppAiContext): String
+
+    suspend fun generate(context: AppAiContext): AppAiDescription?
+}
