--- core/app-ai-description/build.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    alias(libs.plugins.apkanalyzer.library)
-    alias(libs.plugins.apkanalyzer.hilt)
-    alias(libs.plugins.apkanalyzer.room)
-}
-
-android {
-    namespace = "sk.styk.martin.apkanalyzer.core.appaidescription"
-}
-
-dependencies {
-    implementation(projects.core.common)
-    implementation(projects.core.apps)
-    implementation(libs.mlkit.genai.prompt)
-}
