--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/AppAiDescription.kt
@@ -1,3 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription
-
-data class AppAiDescription(val shortDescription: String, val longDescription: String)
