--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/AppAiDescriptionRepository.kt
@@ -1,7 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription
-
-import sk.styk.martin.apkanalyzer.core.common.model.AppReference
-
-interface AppAiDescriptionRepository {
-    suspend fun getDescription(reference: AppReference): AppAiDescription?
-}
