--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/ai/OnDeviceAiEngine.kt
@@ -1,7 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription.ai
-
-internal interface OnDeviceAiEngine {
-    suspend fun isAvailable(): Boolean
-
-    suspend fun runPrompt(prompt: String): String?
-}
