--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/cache/AppAiDescriptionDatabase.kt
@@ -1,9 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription.cache
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(entities = [AppAiDescriptionEntity::class], version = 2, exportSchema = false)
-internal abstract class AppAiDescriptionDatabase : RoomDatabase() {
-    abstract fun appAiDescriptionDao(): AppAiDescriptionDao
-}
