--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/cache/AppAiDescriptionEntity.kt
@@ -1,13 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription.cache
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "app_ai_description")
-internal data class AppAiDescriptionEntity(
-    @PrimaryKey val packageName: String,
-    val inputHash: String,
-    val shortDescription: String,
-    val longDescription: String,
-    val generatedAt: Long,
-)
