--- core/app-ai-description/src/main/kotlin/sk/styk/martin/apkanalyzer/core/appaidescription/metadata/AppMetadataProvider.kt
@@ -1,8 +1,0 @@
-package sk.styk.martin.apkanalyzer.core.appaidescription.metadata
-
-import sk.styk.martin.apkanalyzer.core.appaidescription.AppAiContext
-import sk.styk.martin.apkanalyzer.core.common.model.AppReference
-
-internal interface AppMetadataProvider {
-    suspend fun getAppContext(reference: AppReference): AppAiContext?
-}
