+++ core/ui-library/src/main/kotlin/sk/styk/martin/apkanalyzer/core/uilibrary/components/SkeletonBox.kt
@@ -6,16 +6,21 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import sk.styk.martin.apkanalyzer.core.uilibrary.modifier.shimmer
 import sk.styk.martin.apkanalyzer.core.uilibrary.theme.AppTheme
 
 @Composable
-fun SkeletonBox(modifier: Modifier = Modifier) {
+fun SkeletonBox(
+    modifier: Modifier = Modifier,
+    color: Color = AppTheme.colors.surfaceVariant,
+    highlightColor: Color = AppTheme.colors.surface,
+) {
     Box(
         modifier = modifier
             .clip(RoundedCornerShape(4.dp))
-            .background(AppTheme.colors.surfaceVariant)
-            .shimmer(),
+            .background(color)
+            .shimmer(surfaceColor = color, highlightColor = highlightColor),
     )
 }
