+++ feature/app-detail/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/appdetail/impl/components/aisummary/AiSummaryAction.kt
@@ -1,0 +1,5 @@
+package sk.styk.martin.apkanalyzer.feature.appdetail.impl.components.aisummary
+
+internal sealed interface AiSummaryAction {
+    data object DownloadModel : AiSummaryAction
+}
