+++ feature/app-detail/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/appdetail/impl/components/aisummary/AiSummaryState.kt
@@ -1,0 +1,14 @@
+package sk.styk.martin.apkanalyzer.feature.appdetail.impl.components.aisummary
+
+import androidx.compose.runtime.Immutable
+import sk.styk.martin.apkanalyzer.core.aiinsights.appdescription.AppAiDescription
+
+internal sealed interface AiSummaryState {
+    data object Hidden : AiSummaryState
+    data object Downloadable : AiSummaryState
+    data object Downloading : AiSummaryState
+    data object Loading : AiSummaryState
+
+    @Immutable
+    data class Loaded(val description: AppAiDescription) : AiSummaryState
+}
